+++ src/app/app.component.ts
@@ -1,6 +1,7 @@
 // app.component.js
 import { Component, OnInit } from 'angular2/core';
 import { ROUTER_DIRECTIVES, RouteConfig, ComponentInstruction, CanActivate } from 'angular2/router';
+import { CustomRouterOutlet } from './shared/directives/custom-router-outlet';
 import { checkAuth } from './auth_module/auth/check_auth';
 import { LoginComponent } from './auth_module/login/login.component';
 import { SignupComponent } from './auth_module/signup/signup.component';
@@ -9,8 +10,8 @@ import { HomeComponent } from './home/home.component';
 
 @Component({
     selector: 'app-container',
-    template: '<router-outlet></router-outlet>',
-    directives: [ROUTER_DIRECTIVES],
+    template: '<custom-router-outlet></custom-router-outlet>',
+    directives: [CustomRouterOutlet],
 })
 @RouteConfig([
     { path: '/home', as: 'Home', component: HomeComponent, useAsDefault: true },