+++ src/app/shared/services/data.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from 'angular2/core';
 import { Http, Response } from 'angular2/http';
 import {Observable} from 'rxjs/Observable';
+import {Config} from '../../config/config';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
-
 
 @Injectable()
 export class DataService {
@@ -11,11 +11,11 @@ export class DataService {
     // API path
     baseUrl: string = '/src';
 
-    constructor(private http: Http) { }
+    constructor(private _http: Http, private _config: Config) { }
 
     loginUser(formValues: Object) {
         // change this to your server login api which will be returning token and basic data
-        return this.http.get(this.baseUrl + '/login.json')
+        return this._http.get(this._config.get('apiUrl') + 'login.json')
             .map((res: Response) => res.json())
             .catch(this.handleError);
     }
@@ -23,7 +23,7 @@ export class DataService {
     signupUser(formValues: Object) {
         // change this to your server signup api
         // Will also making user logged in & returning token and basic data
-        return this.http.get(this.baseUrl + '/login.json')
+        return this._http.get(this._config.get('apiUrl') + 'login.json')
             .map((res: Response) => res.json())
             .catch(this.handleError);
     }